refactor(atelier): move event modifier parsing to compat
