+++ crates/vize_atelier_jsx/src/lower/attr.rs
@@ -9,6 +9,8 @@
 
 mod compat;
 
+use compat::split_on_event_modifiers;
+
 use oxc_ast::ast::{
     JSXAttribute, JSXAttributeItem, JSXAttributeName, JSXAttributeValue, JSXSpreadAttribute,
 };
@@ -303,52 +305,4 @@ impl<'a, 'm, 's> Lowerer<'a, 'm, 's> {
             JSXAttributeValue::Fragment(fragment) => Some(self.dyn_expr(fragment.span())),
         }
     }
-}
-
-/// Split a babel-plugin-jsx event attribute name into its event name and
-/// trailing option modifiers, e.g. `onClickCapture` -> `("click", ["capture"])`
-/// and `onInputPassiveCapture` -> `("input", ["passive", "capture"])`.
-///
-/// Returns `None` for names without an `on<Event>` shape, without any
-/// recognized trailing modifier, or whose only content is modifiers (so bare
-/// `onCapture` / `onOnce` keep their plain-bind behavior).
-fn split_on_event_modifiers(name: &str) -> Option<(String, std::vec::Vec<&str>)> {
-    // Require an `on` prefix immediately followed by an uppercase event char.
-    let rest = name.strip_prefix("on")?;
-    if !rest.chars().next()?.is_ascii_uppercase() {
-        return None;
-    }
-
-    // Peel recognized option modifiers off the END, preserving source order.
-    let mut event = rest;
-    let mut mods = std::vec::Vec::new();
-    loop {
-        let modifier = if let Some(head) = event.strip_suffix("Capture") {
-            event = head;
-            "capture"
-        } else if let Some(head) = event.strip_suffix("Once") {
-            event = head;
-            "once"
-        } else if let Some(head) = event.strip_suffix("Passive") {
-            event = head;
-            "passive"
-        } else {
-            break;
-        };
-        mods.push(modifier);
-    }
-    mods.reverse();
-
-    // Require at least one modifier and a non-empty event tail.
-    if mods.is_empty() || event.is_empty() {
-        return None;
-    }
-
-    // Lowercase the first char of the remaining event name.
-    let mut chars = event.chars();
-    let first = chars.next()?;
-    let mut lowered = String::new("");
-    lowered.push(first.to_ascii_lowercase());
-    lowered.push_str(chars.as_str());
-    Some((lowered, mods))
 }
